task 1: **Task 1: Create `card/` package, move `CardBuilder` and templates, wire
